move MetaStore struct to own file
