+++ src/metastore/mod.rs
@@ -2,6 +2,7 @@ mod block;
 mod bucket_meta;
 mod constants;
 mod errors;
+mod meta_store;
 mod object;
 mod stores;
 mod traits;
@@ -10,6 +11,7 @@ pub use block::{Block, BlockID, BLOCKID_SIZE};
 pub use bucket_meta::BucketMeta;
 pub use constants::*;
 pub use errors::{FsError, MetaError};
+pub use meta_store::*;
 pub use object::{Object, ObjectData};
 pub use stores::{FjallStore, FjallStoreNotx};
 pub use traits::*;
